correct test validate
